Add test wavs
Add a "content" test file with melody (and some admittedly complex timbre) and a "style" test file that is a bunch of Glitch sound design. If you're not a Glitch EDM artist. The style file probably doesn't make much sense. I'll make sure to add something boring as an additional test in the future.