Add wav comparison function
Make sure that the input wavs have the same bit depth and sample rate. I'm not going to worry about converting the "style" wav to have the same parameters as the "content" wav at this time. But I could implement that in the future. TODO: Also check that the number of channels are the same across the wavs.