Shift from wavio to librosa
Eliminates need to be as cognizant about input wav parameters when loading wavs.

The librosa load is capable of samplerate conversion during the loading processes; so, it will be easy to ensure that sample rates are the same.

Wavio also had the issue where it couldn't load 32 bit float files.