Tag gram_matrix for further tweaking and testing
Gram Matrix implementation may not be correct. TODO: Check that this version of the gram matrix works with the way the data is handled by this script.

In addition. This commit updates all the transparent network layers to properly return and detach the input tensors passed to them. There was a mismatch in class variable names.